api/requirements.txt
beautifulsoup4==4.9.1
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
firebase-admin==4.3.0
Flask==1.1.2
Flask-PyMongo==2.3.0
future==0.18.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.16.0
lxml==4.5.1
MarkupSafe==1.1.1
msgpack==1.0.0
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
python-dotenv==0.13.0
pytz==2020.1
pyzbar==0.1.8
requests==2.23.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wikipedia==1.4.0
uwsgi==2.0.19.1
flask-cors==3.0.7
